Move Kotlin code to other class
